Re-arranging and template breakout
includes folder now has a folder inside of it called
“template_elements”. This is where the common elements for the info
pages go. I’ve also broken out the CSS files, there are three. One for
the info pages and “whatsnew” page, and one for the splash page.

Please work within the CSS files moving forward. And please work with
the templates for the footer, header, and common elements.